mv interactive terminal viewer to own header file for reuse.
